rename: StoreAccessDeniedException 패키지 이동
- 도메인 안 패키지에서 common.exception 패키지로 이동
