+++ src/main/java/com/example/jjangushrine/exception/common/StoreAccessDeniedException.java
@@ -1,7 +1,6 @@
-package com.example.jjangushrine.domain.product.exception;
+package com.example.jjangushrine.exception.common;
 
 import com.example.jjangushrine.exception.ErrorCode;
-import com.example.jjangushrine.exception.common.BusinessException;
 
 // Store 에 해당 Exception 추가 시 삭제 예정
 public class StoreAccessDeniedException extends BusinessException {
